fix: fix errors from linters
