+++ base_decorator.rb
@@ -1,0 +1,12 @@
+require './nameable'
+
+class BaseDecorator < Nameable
+  def initialize(nameable)
+    @nameable = nameable
+    super
+  end
+
+  def correct_name
+    @nameable.correct_name
+  end
+end
