--- lib/base_decorator.rb
@@ -1,11 +1,0 @@
-require "./nameable"
-
-class BaseDecorator < Nameable
-    def initialize(nameable)
-        @nameable = nameable
-    end
-
-    def correct_name
-        @nameable.correct_name
-    end
-end
